Bump the SHA-256 sum of the Dangerzone MSI
We have rebuilt the Dangerzone MSI for Windows using WiX toolset v3.14.
Originally, we had built it with v3.11, but we wanted to update to the
latest version for good measure. The Dangerzone application has remained
exactly the same, but the contents of the MSI file have slightly
changed, due to the new WiX toolset. As a result, we have changed the
name from `Dangerzone-0.6.0.msi` to `Dangerzone-0.6.0-1.msi`.