Add link to Tails website
Point users to the installation instructions of Dangerzone in the Tails
website. These instructions were recently added to Tails, and we have
worked with the Tails developers to make this integration happen.

As for the Tails logo, we have downloaded it from the promotional
material section of the Tails website:
https://tails.net/contribute/how/promote/material/logo/tails-logo-drawing.svg

See also:
* https://tails.net/news/dangerzone/index.en.html
* https://gitlab.tails.boum.org/tails/tails/-/issues/20355