about: Explain that this section is not the original one
Explain that the current "About" section is not the original article
that Micah wrote when he released Dangerzone. Since then, Dangerzone has
undergone some architectural changes, which must be reflected in this
article. So, in order to clear any confusion, it's best to add that the
original article was written by Micah, and then the Dangerzone
contributors have updated it.